refactor : Cars - String 배열로 생성하는 로직 CarsFactory로 분리
