+++ src/main/java/racingcar/game/view/InputView.java
@@ -2,6 +2,7 @@ package racingcar.game.view;
 
 import camp.nextstep.edu.missionutils.Console;
 import racingcar.domain.cars.Cars;
+import racingcar.domain.cars.CarsFactory;
 import racingcar.domain.race.RaceCount;
 
 public class InputView {
@@ -14,6 +15,6 @@ public class InputView {
     public Cars getCars() {
         String readLine = Console.readLine();
 
-        return Cars.fromNames(readLine.split(","));
+        return CarsFactory.fromNames(readLine.split(","));
     }
 }
